requirements.txt
torch>=2.0.0
torchvision>=0.15.0
omegaconf>=2.3.0
opencv-python>=4.7.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.2.0
numpy~=2.2.4
pillow~=11.1.0
mpmath~=1.3.0